[Java][breaking] move package namespace to "com.novi"
