+++ serde-generate/runtime/java/com/novi/bincode/BincodeSerializer.java
@@ -1,10 +1,10 @@
 // Copyright (c) Facebook, Inc. and its affiliates
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
-package com.facebook.bincode;
+package com.novi.bincode;
 
 import java.lang.Exception;
-import com.facebook.serde.BinarySerializer;
+import com.novi.serde.BinarySerializer;
 
 public class BincodeSerializer extends BinarySerializer {
     public void serialize_len(long value) throws Exception {
