+++ serde-generate/runtime/java/com/novi/lcs/LcsSerializer.java
@@ -1,11 +1,11 @@
 // Copyright (c) Facebook, Inc. and its affiliates
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
-package com.facebook.lcs;
+package com.novi.lcs;
 
 import java.lang.Exception;
-import com.facebook.serde.Slice;
-import com.facebook.serde.BinarySerializer;
+import com.novi.serde.Slice;
+import com.novi.serde.BinarySerializer;
 
 public class LcsSerializer extends BinarySerializer {
     private void serialize_u32_as_uleb128(int value) {
