+++ serde-generate/runtime/java/com/novi/lcs/LcsTest.java
@@ -1,7 +1,7 @@
 // Copyright (c) Facebook, Inc. and its affiliates
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
-package com.facebook.lcs;
+package com.novi.lcs;
 
 import java.util.Arrays;
 import java.lang.reflect.Method;
