+++ serde-generate/runtime/java/com/novi/serde/Serializer.java
@@ -1,7 +1,7 @@
 // Copyright (c) Facebook, Inc. and its affiliates
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
-package com.facebook.serde;
+package com.novi.serde;
 
 import java.lang.Exception;
 import java.math.BigInteger;
