+++ serde-generate/runtime/java/com/novi/serde/Unit.java
@@ -1,7 +1,7 @@
 // Copyright (c) Facebook, Inc. and its affiliates
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
-package com.facebook.serde;
+package com.novi.serde;
 
 public final class Unit {
     public Unit() {}
