+++ serde-generate/runtime/java/com/novi/serde/Unsigned.java
@@ -1,12 +1,11 @@
 // Copyright (c) Facebook, Inc. and its affiliates
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
-package com.facebook.serde;
+package com.novi.serde;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Target;
 
 @Target({ElementType.TYPE_USE})
-public @interface ArrayLen {
-    int length();
+public @interface Unsigned {
 }
